pyproject.toml
[tool.poetry]
name = "chack-check users service"
version = "0.1.0"
description = "Users service for Chack-Check project"
authors = ["artemowkin <artyom.groshev2017@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
strawberry-graphql = {extras = ["fastapi"], version = "^0.194.4"}
uvicorn = "^0.22.0"
email-validator = "^2.0.0.post2"
sqlalchemy = "^2.0.18"
pydantic-settings = "^2.0.1"
redis = "^4.6.0"
psycopg-binary = "^3.1.9"
psycopg = "^3.1.9"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"


[tool.poetry.group.lint.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
